requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
cssutils==1.0.2
Django==3.1.6
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
lxml==4.6.2
phonenumbers==8.12.18
premailer==3.7.0
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
yagmail==0.14.245
